work on review changes and some refactoring
