+++ .github/CODEOWNERS
@@ -388,8 +388,6 @@
 
 # Speed Improvement team.
 /app.yaml @vojtechjelinek
-/core/controllers/analytics.py @jamesjay4199
-/core/controllers/analytics_test.py @jamesjay4199
 /core/templates/dev/head/pages/Base.ts @vojtechjelinek
 /core/templates/dev/head/pages/base.html @vojtechjelinek
 /core/templates/dev/head/pages/footer.html @vojtechjelinek
