+++ core/controllers/base_test.py
@@ -888,3 +888,19 @@ class SignUpTests(test_utils.GenericTestBase):
         )
 
         self.get_html_response('/about')
+
+
+class GoogleAnalyticsHandlerTests(test_utils.GenericTestBase):
+
+    def test_analytics_handler_returns_required_values(self):
+
+        values_dict = {
+            'analytics_id': 'test123',
+            'site_name_for_analytics': 'Oppia foundation'
+        }
+
+        with self.swap(feconf, 'ANALYTICS_ID', 'test123'):
+            with self.swap(
+                feconf, 'SITE_NAME_FOR_ANALYTICS', 'Oppia foundation'):
+                response = self.get_json(feconf.GOOGLE_ANALYTICS_DATA_URL)
+                self.assertEqual(response, values_dict)
