+++ feconf.py
@@ -577,7 +577,7 @@ TASK_URL_SUGGESTION_EMAILS = (
 # TODO(sll): Add all other URLs here.
 ADMIN_URL = '/admin'
 ADMIN_ROLE_HANDLER_URL = '/adminrolehandler'
-ANALYTICS_DATA_URL = '/analytics'
+GOOGLE_ANALYTICS_DATA_URL = '/analytics'
 COLLECTION_DATA_URL_PREFIX = '/collection_handler/data'
 COLLECTION_EDITOR_DATA_URL_PREFIX = '/collection_editor_handler/data'
 COLLECTION_SUMMARIES_DATA_URL = '/collectionsummarieshandler/data'
@@ -954,6 +954,6 @@ AVAILABLE_LANDING_PAGES = {
     'maths': ['fractions', 'ratios']
 }
 
-# Data required for google analytics.
+# Data required for Google Analytics.
 ANALYTICS_ID = ''
 SITE_NAME_FOR_ANALYTICS = ''
