+++ main.py
@@ -20,7 +20,6 @@ import logging
 from constants import constants
 from core.controllers import acl_decorators
 from core.controllers import admin
-from core.controllers import analytics
 from core.controllers import base
 from core.controllers import classifier
 from core.controllers import collection_editor
@@ -667,8 +666,8 @@ URLS = MAPREDUCE_HANDLERS + [
         r'/resolveissuehandler/<exploration_id>', editor.ResolveIssueHandler),
 
     get_redirect_route(
-        r'%s' % feconf.ANALYTICS_DATA_URL,
-        analytics.AnalyticsHandler),
+        r'%s' % feconf.GOOGLE_ANALYTICS_DATA_URL,
+        base.GoogleAnalyticsHandler),
 
     # 404 error handler.
     get_redirect_route(r'/<:.*>', base.Error404Handler),
